requirements.txt
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.2
borb==2.1.8
certifi==2022.12.7
charset-normalizer==3.0.1
exceptiongroup==1.1.0
fake-useragent==1.1.1
fonttools==4.38.0
h11==0.14.0
idna==3.4
outcome==1.2.0
Pillow==9.4.0
pypdf==3.3.0
PyPDF2==3.0.1
PyPDF4==1.27.0
pypng==0.20220715.0
PySocks==1.7.1
python-barcode==0.14.0
qrcode==7.4
requests==2.28.2
selenium==4.8.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.14
wsproto==1.2.0
